+++ src/Eloquent/CollectionPaginator.php
@@ -5,6 +5,7 @@ namespace MallardDuck\LaravelTraits\Eloquent;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Facades\Request;
 
 /**
  * Allows you to utilize the Eloquent\Builder's paginate methods on regular Eloquent Collections.
@@ -18,16 +19,22 @@ trait CollectionPaginator
     public function bootCollectionPaginator()
     {
         Collection::macro('paginate', function ($perPage = 5, $pageName = 'page', $page = null) {
+            /**
+             * @var Collection $this
+             */
             $total = $this->count();
             $page = $page ?: Paginator::resolveCurrentPage($pageName);
 
             $results = $this->slice(($page - 1) * $perPage, $perPage)->all();
 
-            $currentQuery = app('request')->query();
+            /**
+             * @var Request $currentQuery
+             */
+            $currentRequest = app('request');
 
             return new LengthAwarePaginator($results, $total, $perPage, $page, [
             'path' => Paginator::resolveCurrentPath(),
-            'query' => app('request')->query(),
+            'query' => $currentRequest->query(),
             'pageName' => $pageName,
             ]);
         });
@@ -37,11 +44,14 @@ trait CollectionPaginator
 
             $results = $this->slice(($page - 1) * $perPage, $perPage + 1)->all();
 
-            $currentQuery = app('request')->query();
+            /**
+             * @var Request $currentQuery
+             */
+            $currentRequest = app('request');
 
             return new Paginator($results, $perPage, $page, [
             'path' => Paginator::resolveCurrentPath(),
-            'query' => app('request')->query(),
+            'query' => $currentRequest->query(),
             'pageName' => $pageName,
             ]);
         });